requirements.txt
flask
mxnet
gluoncv
matplotlib
Pillow